Refactor analysis for defun
