+++ src/analyse.h
@@ -5,4 +5,5 @@
 typedef struct _Project Project;
 
 void analyse_ast(Project *project, Node *root);
+void analyse_node(Project *project, Node *node, gchar **context);
