+++ src/analyse_defpackage.c
@@ -1,4 +1,4 @@
-#include "defpackage.h"
+#include "analyse_defpackage.h"
 #include "package.h"
 #include <string.h>
