+++ src/analyse_defun.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "project.h"
+#include "node.h"
+
+void analyse_defun(Project *project, Node *expr, gchar **context);
+
