+++ tests/analyser_test.c
@@ -1,4 +1,4 @@
-#include "analyser.h"
+#include "analyse.h"
 #include "lisp_lexer.h"
 #include "lisp_parser.h"
 #include "node.h"
